Move AllocatedData in VectorListBuffer into auxiliary data
