+++ src/common/vector/list_vector.cpp
@@ -2,6 +2,83 @@
 
 namespace duckdb {
 
+VectorListBuffer::VectorListBuffer(Allocator &allocator, idx_t capacity, unique_ptr<Vector> vector,
+                                   idx_t child_capacity)
+    : VectorBuffer(VectorBufferType::LIST_BUFFER), child(std::move(vector)), capacity(child_capacity) {
+	if (capacity > 0) {
+		auto allocated_data = allocator.Allocate(capacity * sizeof(list_entry_t));
+		data_ptr = allocated_data.get();
+		AddAuxiliaryData(make_uniq<AuxiliaryAllocatedDataHolder>(std::move(allocated_data)));
+	}
+}
+VectorListBuffer::VectorListBuffer(Allocator &allocator, idx_t capacity, const LogicalType &list_type,
+                                   idx_t child_capacity)
+    : VectorListBuffer(allocator, capacity, make_uniq<Vector>(ListType::GetChildType(list_type), child_capacity),
+                       child_capacity) {
+}
+
+VectorListBuffer::VectorListBuffer(idx_t capacity, const LogicalType &list_type, idx_t child_capacity)
+    : VectorListBuffer(Allocator::DefaultAllocator(), capacity, list_type, child_capacity) {
+}
+
+VectorListBuffer::VectorListBuffer(data_ptr_t data, const VectorListBuffer &parent)
+    : VectorBuffer(VectorBufferType::LIST_BUFFER), data_ptr(data), capacity(parent.capacity), size(parent.size) {
+	child = make_uniq<Vector>(Vector::Ref(parent.GetChild()));
+}
+
+VectorListBuffer::VectorListBuffer(AllocatedData allocated_data_p, const VectorListBuffer &parent)
+    : VectorBuffer(VectorBufferType::LIST_BUFFER), capacity(parent.capacity), size(parent.size) {
+	child = make_uniq<Vector>(Vector::Ref(parent.GetChild()));
+	data_ptr = allocated_data_p.get();
+	AddAuxiliaryData(make_uniq<AuxiliaryAllocatedDataHolder>(std::move(allocated_data_p)));
+}
+
+void VectorListBuffer::Reserve(idx_t to_reserve) {
+	if (to_reserve > capacity) {
+		if (to_reserve > DConstants::MAX_VECTOR_SIZE) {
+			// overflow: throw an exception
+			throw OutOfRangeException("Cannot resize vector to %d rows: maximum allowed vector size is %s", to_reserve,
+			                          StringUtil::BytesToHumanReadableString(DConstants::MAX_VECTOR_SIZE));
+		}
+		idx_t new_capacity = NextPowerOfTwo(to_reserve);
+		D_ASSERT(new_capacity >= to_reserve);
+		child->Resize(capacity, new_capacity);
+		capacity = new_capacity;
+	}
+}
+
+void VectorListBuffer::Append(const Vector &to_append, idx_t to_append_size, idx_t source_offset) {
+	Reserve(size + to_append_size - source_offset);
+	VectorOperations::Copy(to_append, *child, to_append_size, source_offset, size);
+	size += to_append_size - source_offset;
+}
+
+void VectorListBuffer::Append(const Vector &to_append, const SelectionVector &sel, idx_t to_append_size,
+                              idx_t source_offset) {
+	Reserve(size + to_append_size - source_offset);
+	VectorOperations::Copy(to_append, *child, sel, to_append_size, source_offset, size);
+	size += to_append_size - source_offset;
+}
+
+void VectorListBuffer::PushBack(const Value &insert) {
+	while (size + 1 > capacity) {
+		child->Resize(capacity, capacity * 2);
+		capacity *= 2;
+	}
+	child->SetValue(size++, insert);
+}
+
+void VectorListBuffer::SetCapacity(idx_t new_capacity) {
+	this->capacity = new_capacity;
+}
+
+void VectorListBuffer::SetSize(idx_t new_size) {
+	this->size = new_size;
+}
+
+VectorListBuffer::~VectorListBuffer() {
+}
+
 template <class T>
 T &ListVector::GetEntryInternal(T &vector) {
 	D_ASSERT(vector.GetType().id() == LogicalTypeId::LIST || vector.GetType().id() == LogicalTypeId::MAP);
